pyproject.toml
[project]
name = "srag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # Core LLM & RAG
    "langchain>=1.0.8",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.0",
    
    # Embeddings
    "sentence-transformers>=5.1.2",
    
    # Vector Databases
    "chromadb>=1.3.5",
    "faiss-cpu>=1.13.0",
    
    # Document Processing
    "pymupdf>=1.26.6",
    "pypdf>=6.4.0",
    
    # LLM Providers
    "google-genai>=1.56.0",
    
    # Configuration & Environment
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    
    # Utilities
    "tqdm>=4.67.1",
]
